requirements.txt
zcomcule-ml-commons
pyyaml
azure-ai-anomalydetector
google-auth
google-cloud-storage
plotly
kaleido
matplotlib